IntExt_Trigger
I don't know why this is not working?? 
I get a compiler error saying the referenced script (IntExt_Trigger) on this Behaviour is missing. 

I did some digging on StackOverflow and the unity forums and it apparently has something to do with the filename not matching the public class set up inside the script but... weird thing is... It does!! 
Unless I'm missing something I don't know what I'm doing wrong here.